Hide the shields stats/controls when shields are off for a specific site
- Also hides the brave icon when no URL is loaded
- Fixes the height of a popover by taking the popover arrow into account
